Swap DailyReward and DailyReward2
