+++ .Lib9c.Tests/BlockPolicyTest.cs
@@ -127,12 +127,12 @@ namespace Lib9c.Tests
 
             var singleAction = new PolymorphicAction<ActionBase>[]
             {
-                new DailyReward2(),
+                new DailyReward(),
             };
             var manyActions = new PolymorphicAction<ActionBase>[]
             {
-                new DailyReward2(),
-                new DailyReward2(),
+                new DailyReward(),
+                new DailyReward(),
             };
             Transaction<PolymorphicAction<ActionBase>> txWithSingleAction =
                 Transaction<PolymorphicAction<ActionBase>>.Create(
@@ -289,7 +289,7 @@ namespace Lib9c.Tests
 
             blockChain.MakeTransaction(
                 adminPrivateKey,
-                new PolymorphicAction<ActionBase>[] { new DailyReward2(), }
+                new PolymorphicAction<ActionBase>[] { new DailyReward(), }
             );
 
             await blockChain.MineBlock(adminAddress);
