+++ Lib9c/Action/DailyReward2.cs
@@ -1,19 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Linq;
 using Bencodex.Types;
 using Libplanet;
 using Libplanet.Action;
-using Nekoyume.Model.Item;
-using Nekoyume.Model.Mail;
 using Nekoyume.Model.State;
-using Nekoyume.TableData;
 
 namespace Nekoyume.Action
 {
     [Serializable]
-    [ActionType("daily_reward2")]
+    [ActionType("daily_reward")]
     public class DailyReward2 : GameAction
     {
         public Address avatarAddress;
@@ -46,25 +42,6 @@ namespace Nekoyume.Action
                 avatarState.actionPoint = gameConfigState.ActionPointMax;
             }
 
-            // create item
-            var materialSheet = states.GetSheet<MaterialItemSheet>();
-            var materials = new Dictionary<Material, int>();
-            var material = ItemFactory.CreateMaterial(materialSheet, 400000);
-            materials[material] = 10;
-
-            var result = new DailyRewardResult
-            {
-                materials = materials,
-            };
-            
-            // create mail
-            var mail = new DailyRewardMail(result, 
-                                           ctx.BlockIndex, 
-                                           ctx.Random.GenerateRandomGuid(),
-                                           ctx.BlockIndex);
-
-            result.id = mail.id;
-            avatarState.UpdateV3(mail);
             return states.SetState(avatarAddress, avatarState.Serialize());
         }
 
@@ -76,35 +53,6 @@ namespace Nekoyume.Action
         protected override void LoadPlainValueInternal(IImmutableDictionary<string, IValue> plainValue)
         {
             avatarAddress = plainValue["avatarAddress"].ToAddress();
-        }
-        
-        
-        [Serializable]
-        public class DailyRewardResult : AttachmentActionResult
-        {
-            public Dictionary<Material, int> materials;
-            public Guid id;
-
-            protected override string TypeId => "dailyReward.dailyRewardResult";
-
-            public DailyRewardResult()
-            {
-            }
-
-            public DailyRewardResult(Bencodex.Types.Dictionary serialized) : base(serialized)
-            {
-                materials = serialized["materials"].ToDictionary_Material_int();
-                id = serialized["id"].ToGuid();
-            }
-
-            public override IValue Serialize() =>
-#pragma warning disable LAA1002
-                new Bencodex.Types.Dictionary(new Dictionary<IKey, IValue>
-                {
-                    [(Text) "materials"] = materials.Serialize(),
-                    [(Text) "id"] = id.Serialize(),
-                }.Union((Bencodex.Types.Dictionary) base.Serialize()));
-#pragma warning restore LAA1002
         }
     }
 }
